radicale: Do not bother to rotate the log file
Without running radicale in debug mode, the log file is quite small, so
do not bother to rotate it.  Actually, it's hard to configure uwsgi to
properly rotate the log file: when the uwsgi worker accepts the 'HUP'
signal and restart, it doesn't have priviledge to create the socket file
and thus failed.

If, on the other hand, the uwsgi isn't stopped properly, it will leave
SysV semaphores in use, and finally exaust all semaphores.